requirements.txt
python-dotenv~=1.1.1
pymongo~=4.14.1
fastapi~=0.116.1
uvicorn~=0.35.0
kafka-python
scikit-learn
